requirements.txt
Django==1.11.3
psycopg2
pytz==2017.2
django-tinymce
